requirements.txt
pytesseract
pillow
opencv-python
